:memo: The `actual` method was unnaturally split, so I revised it to match the definition of the `expect` method.
