--- core/droidkaigiui/src/androidMain/kotlin/io/github/droidkaigi/confsched/droidkaigiui/extension/EnableMouseDragScroll.android.kt
@@ -1,8 +1,0 @@
-package io.github.droidkaigi.confsched.droidkaigiui.extension
-
-import androidx.compose.foundation.gestures.ScrollableState
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-
-@Composable
-actual fun Modifier.enableMouseDragScroll(scrollableState: ScrollableState): Modifier = this // NOOP
