+++ core/droidkaigiui/src/androidMain/kotlin/io/github/droidkaigi/confsched/droidkaigiui/extension/ModifierExtensions.android.kt
@@ -1,7 +1,11 @@
 package io.github.droidkaigi.confsched.droidkaigiui.extension
 
+import androidx.compose.foundation.gestures.ScrollableState
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
+
+@Composable
+actual fun Modifier.enableMouseDragScroll(scrollableState: ScrollableState): Modifier = this // NOOP
 
 @Composable
 actual fun Modifier.bindMouseBackForward(
